sistema_integracion/backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==24.3
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
qrcode==7.4.2
twilio==8.10.0
python-decouple==3.8
pymysql==1.1.2
